:adhesive_bandage: (Authentication-api): fixed code structure
