+++ libs/api/authentication/api/feature/src/lib/api-authentication-api-feature.module.ts
@@ -3,18 +3,18 @@ import { ApiAuthenticationResolver } from './api-authentication-api.resolver';
 import { forwardRef, Module } from '@nestjs/common';
 import { UsersModule } from './users/users.module';
 import { AuthModule } from './auth/auth.module';
-import { UsersService } from './api-authentication-api.service';
+import { AuthService } from '@graduates/api/authentication/service/feature';
 // import { UsersService } from '@graduates/api/authentication/service/feature';
 //import { UsersService } from '@graduates/api/authentication/service/feature';
 //import { ApiAuthenticationApiSharedInterfacesDataAccessModule } from '@graduates/api/authentication/service/feature';
 
 
 @Module({
-  imports: [UsersModule, AuthModule, UsersService
+  imports: [UsersModule, AuthModule, AuthService
     //ApiAuthenticationApiSharedInterfacesDataAccessModule
   ],
   controllers: [],
-  providers: [ ApiAuthenticationResolver, UsersService 
+  providers: [ ApiAuthenticationResolver, AuthService 
   ],
 })
 export class ApiAuthenticationApiFeatureModule {}
