+++ libs/api/authentication/api/feature/src/lib/auth/auth.module.ts
@@ -1,11 +1,9 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
 import { AuthResolver } from './auth.resolver';
-import { LocalStrategy } from './local.strategy';
 import { PassportModule } from '@nestjs/passport'
 import { UsersModule } from '../users/users.module';
 import { JwtModule } from '@nestjs/jwt';
-import { JwtStrategy } from './jwt-strategy';
+import { JwtStrategy, LocalStrategy, AuthService } from '@graduates/api/authentication/service/feature';
 //import { GoogleStrategy } from '@graduates/api/authentication/service/feature';
 
 @Module({
