+++ libs/api/authentication/api/feature/src/lib/users/users.module.ts
@@ -1,10 +1,10 @@
-import { UsersService } from '../api-authentication-api.service';
+import { AuthService } from '@graduates/api/authentication/service/feature';
 import { Module } from '@nestjs/common';
 import { UsersResolver } from './users.resolver';
 
 @Module({
-  providers: [UsersService, UsersResolver
+  providers: [AuthService, UsersResolver
   ],
-  exports: [UsersService]
+  exports: [AuthService]
 })
 export class UsersModule {}
