+++ libs/api/authentication/service/feature/src/index.ts
@@ -3,3 +3,9 @@ export * from './lib/strategy/google.strategy'
 //export * from './lib/api-authentication-api-service.module'
 export * from './lib/commands/RgisterCommand'
 export * from './lib/queries/LoginQuery'
+export * from './lib/queries/login-user-input'
+export * from './lib/queries/login-user-input'
+export * from './lib/strategy/jwt-strategy'
+export * from './lib/strategy/local.strategy'
+export * from './lib/response/login-response'
+export * from './lib/api-authentication-api-service.service'
