+++ libs/api/authentication/service/feature/src/lib/strategy/jwt-strategy.ts
