+++ libs/api/authentication/service/feature/src/lib/strategy/local.strategy.ts
@@ -1,7 +1,7 @@
 import { Strategy } from 'passport-local';
 import {PassportStrategy} from '@nestjs/passport';
 import {Injectable, UnauthorizedException} from '@nestjs/common';
-import { AuthService } from './auth.service';
+import { AuthService } from '../api-authentication-api-service.service';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
